Align retrieval service with design spec: multi-channel RRF, strict public schema, lifecycle fixes
Port multi-channel lexical retrieval (content/path/term channels) with RRF
fusion from knowhere-kb reference. Run graph and lexical channels in parallel
with graceful fallback. Strip internal fields from public response using an
allowlist. Return HTTP 404 for missing documents, make archive idempotent,
fix graph edge weight type to Float, and sort connected chunks by sort_order.

+++ apps/api/alembic/versions/c3d4e5f6a7b8_add_retrieval_service_v1.py
@@ -147,7 +147,7 @@ def upgrade() -> None:
     sa.Column('owner_document_id', sa.String(length=36), nullable=False),
     sa.Column('job_result_id', sa.String(length=36), nullable=False),
     sa.Column('is_directed', sa.Boolean(), nullable=False),
-    sa.Column('weight', sa.Integer(), nullable=True),
+    sa.Column('weight', sa.Float(), nullable=True),
     sa.Column('properties', sa.JSON(), nullable=True),
     sa.Column('created_at', sa.DateTime(), nullable=False),
     sa.Column('updated_at', sa.DateTime(), nullable=False),
